python_elinity-main/pyproject.toml
[project]
name = "python-elinity"
version = "0.1.0"
description = "AI-powered social connection"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=0.7.0",
    "alembic>=1.7.3",
    "assemblyai>=0.40.2",
    "bcrypt==4.0.1",
    "boto3>=1.38.16",
    "celery[redis]>=5.5.2",
    "chromadb>=0.4.18",
    "email-validator>=1.1.3",
    "faker>=20.0.0",
    "fastapi-csrf-protect>=0.1.0",
    "fastapi[standard]>=0.68.0",
    "firebase-admin>=6.8.0",
    "flask>=2.0.0",
    "flower>=2.0.1",
    "google-cloud-storage>=3.1.0",
    "google-genai>=1.15.0",
    "google-generativeai>=0.3.1",
    "gtts>=2.5.4",
    "httpx>=0.19.0",
    "itsdangerous>=2.0.1",
    "jinja2>=3.0.1",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-google-genai>=2.0.10",
    "langgraph>=0.4.7",
    "numpy>=1.23.0",
    "pandas>=2.2.3",
    "passlib[bcrypt]>=1.7.4",
    "pinecone>=7.0.1",
    "psycopg2-binary>=2.9.5",
    "pydantic>=2.5.1",
    "pydantic-settings>=2.0.0",
    "pymilvus[model]>=2.5.10",
    "pymongo>=4.13.0",
    "python-dateutil>=2.8.2",
    "python-dotenv>=0.19.0",
    "python-jose[cryptography]==3.3.0",
    "python-multipart>=0.0.5",
    "python-slugify>=5.0.2",
    "pytz>=2021.3",
    "qdrant-client>=1.1.1",
    "redis>=4.0.1",
    "requests>=2.28.0",
    "sentence-transformers>=4.1.0",
    "speechrecognition>=3.10.0",
    "sqlalchemy>=1.4.48",
    "streamlit>=1.41.0",
    "streamlit-aggrid>=0.2.0",
    "streamlit-authenticator>=0.1.0",
    "uvicorn>=0.15.0",
    "websockets>=15.0.1",
]
